fixing structure all system

+++ app/Services/Inventory/ItemServices.php
@@ -7,6 +7,25 @@ use App\Models\Inventory\ItemVariant;
 use App\Models\Inventory\ItemVariantAttribute;
 
 class ItemServices {
+  /**
+   * Summary of getSku
+   * @param \App\Models\Inventory\ItemVariant|\App\Models\Inventory\Item $item
+   */
+  public static function getSku(mixed $item, array|null $attributes = null) {
+
+    if ($item->format_variant === null || $item->format_variant === '') {
+      return $item->item_code;
+    }
+    $attributesValues = $item instanceof Item ? $attributes : $item->values->toArray();
+    $data = array_combine(array_column($attributesValues, 'attribute_id'), array_column($attributesValues, 'value'));
+    $data['item'] = $item instanceof Item ? $item->code : ($item->item_code ?? $item->item->code);
+    $result = preg_replace_callback('/@\[(.*?)\]\((.*?)\)/', function ($matches) use ($data) {
+      $key = $matches[1];
+      $id = $matches[2];
+      return $data[$key] ?? ($data[$id] ?? $matches[0]);
+    }, $item->format_variant ?? $item->item->format_variant);
+    return $result;
+  }
   public function updateUom(Item $item, array $uoms) {
     foreach ($uoms as $uom) {
       if ($uom['isCustom'] ?? false) {
@@ -73,7 +92,11 @@ class ItemServices {
         'item_id' => $item->id,
         'format_variant' => null,
       ], [
+        'code' => $item->code,
         'item_code' => $item->code,
+        'item_name' => $item->name,
+        'category_id' => $item->category_id,
+        'default_unit_id' => $item->default_unit_id,
       ]);
       return $itemVariant;
     }
@@ -93,8 +116,12 @@ class ItemServices {
       if (!$variant) {
         $variantId = ItemVariant::create([
           'item_id' => $item->id,
+          'code' => static::getSku($item, $prefix),
           'item_code' => $item->code,
+          'item_name' => $item->name,
           'format_variant' => $item->format_variant,
+          'category_id' => $item->category_id,
+          'default_unit_id' => $item->default_unit_id,
         ])->id;
         foreach ($prefix as $attribute) {
           ItemVariantAttribute::create([
@@ -106,7 +133,11 @@ class ItemServices {
       } else {
         $variantId = $variant->update([
           'format_variant' => $item->format_variant,
+          'code' => static::getSku($item, $prefix),
           'item_code' => $item->code,
+          'item_name' => $item->name,
+          'category_id' => $item->category_id,
+          'default_unit_id' => $item->default_unit_id,
         ]);
         $variantId = $variant->id;
       }
@@ -139,7 +170,8 @@ class ItemServices {
     }
   }
 
-  public function updateBarcodes(ItemVariant $variant, array $barcodes) {
+  public function updateBarcodes(ItemVariant|null $variant, array $barcodes) {
+    if (!$variant) return;
     $variant->barcodes()->delete();
     foreach ($barcodes as $barcode) {
       $variant->barcodes()->create([
